% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcehelp.R
\name{add_var}
\alias{add_var}
\title{Create folder skeleton for RCE variable}
\usage{
add_var(variable, category)
}
\arguments{
\item{variable}{the variable to be added}

\item{category}{the type of variable being added, also the parent folder}
}
\description{
Create folder skeleton for RCE variable
}
