% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_files.R
\name{large_unique}
\alias{large_unique}
\title{Get a vector of all unique values in column}
\usage{
large_unique(path, var)
}
\arguments{
\item{path}{path to a csv file representing a dataframe}

\item{var}{the variable to get the identical values for}
}
\description{
Get a vector of all unique values in column
}
