% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcehelp.R
\name{make_readme}
\alias{make_readme}
\title{Create skeleton for RCE readme}
\usage{
make_readme(name = "John Doe",
  description = "Describe purpose of this directory", file_list = T,
  path = ".")
}
\arguments{
\item{name}{the author of the readme file}

\item{file_list}{Boolean, should the directory's files be
listed in the readme?}

\item{path}{the path to the directory that the readme is for}
}
\description{
Create skeleton for RCE readme
}
