% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large_files.R
\name{max.large_freq}
\alias{max.large_freq}
\title{Return the most common value from a large file frequency table}
\usage{
\method{max}{large_freq}(frq, ...)
}
\description{
Return the most common value from a large file frequency table
}
